Moved `Tabulator` and related tests to its own module `crate::tabulator`.
